Use Gulp for JVM build.
Gulp isn't necessary, but Grunt fails in Nashorn for whatever
reason and the piping syntax of Gulp makes it super easy to
build the pipeline from Less to autoprefixer (and any other PostCSS
processor we might want).

There is still build.js, but it doesn't do autoprefixer and was
getting pretty messy.