Create a lite build of ID7
id7.lite.css is a lighter-weight build of ID7.  It excludes styling that
may not be necessary in the context of a mobile/web application.

Excluded styles include the ID7 navigation, footer and content boxes,
ID6 legacy styling, and most Bootstrap components.

In the current configuration (a proof-of-concept, more than anything),
id7.min.css is 185K and id7.lite.min.css is 87K.  There may be scope for
this to decrease further, depending on the app.